# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Buffett's Buffet API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

/**
 * Traffic light indicator
 */
export type MetricScoreColor = typeof MetricScoreColor[keyof typeof MetricScoreColor];


export const MetricScoreColor = {
  green: 'green',
  amber: 'amber',
  red: 'red',
  gray: 'gray',
} as const;

export interface MetricScore {
  /** Machine-readable metric key */
  key: string;
  /** Human-readable metric name */
  name: string;
  /**
     * Raw calculated value
     * @nullable
     */
  value: number | null;
  /**
     * Display-formatted value (e.g. "12.4%", "5.2x")
     * @nullable
     */
  formattedValue?: string | null;
  /**
     * Percentile rank vs S&P500 dataset (0-100)
     * @nullable
     */
  percentile: number | null;
  /** Traffic light indicator */
  color: MetricScoreColor;
  /**
     * One-line explanation of why Buffett cares about this metric
     * @nullable
     */
  explanation: string | null;
  /** Why Buffett weights this metric */
  buffettWhy: string;
  /** True when this metric is excluded for sector reasons */
  notApplicable: boolean;
}

export interface SimilarStock {
  ticker: string;
  name: string;
  /** Pre-calculated Buffett score (0-100) */
  score: number;
  sector: string;
}

/**
 * green=70+, amber=40-69, red=<40
 */
export type StockResultScoreColor = typeof StockResultScoreColor[keyof typeof StockResultScoreColor];


export const StockResultScoreColor = {
  green: 'green',
  amber: 'amber',
  red: 'red',
} as const;

export interface StockResult {
  ticker: string;
  /** @nullable */
  companyName: string | null;
  /** @nullable */
  sector?: string | null;
  /** @nullable */
  filingDate: string | null;
  /** Weighted Buffett score 0-100 */
  buffettScore: number;
  /** green=70+, amber=40-69, red=<40 */
  scoreColor: StockResultScoreColor;
  /** Percentage of applicable metric weight covered by actual data (0-100) */
  scoreConfidence: number;
  metrics: MetricScore[];
  disclaimer: string;
  /** 4 stocks from the scored universe whose Buffett scores are closest to this one */
  similarStocks: SimilarStock[];
}

export type StockSummaryScoreColor = typeof StockSummaryScoreColor[keyof typeof StockSummaryScoreColor];


export const StockSummaryScoreColor = {
  green: 'green',
  amber: 'amber',
  red: 'red',
} as const;

export interface StockSummary {
  ticker: string;
  /** @nullable */
  companyName: string | null;
  /** @nullable */
  sector: string | null;
  buffettScore: number;
  scoreColor: StockSummaryScoreColor;
}