# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Buffett's Buffet API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: stocks
    description: Stock screening operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /stocks/{ticker}:
    get:
      operationId: getStock
      tags: [stocks]
      summary: Get Buffett score for a stock
      description: Fetches live financial data and scores the stock against Buffett's 8 core metrics
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stock data and Buffett score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockResult"
        "404":
          description: Ticker not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Invalid ticker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /stocks/{ticker}/similar:
    get:
      operationId: getSimilarStocks
      tags: [stocks]
      summary: Get similar high-scoring stocks
      description: Returns a curated list of well-known stocks with strong Buffett scores for comparison
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of similar stocks with their scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    MetricScore:
      type: object
      required:
        - key
        - name
        - value
        - percentile
        - color
        - explanation
        - buffettWhy
        - notApplicable
      properties:
        key:
          type: string
          description: Machine-readable metric key
        name:
          type: string
          description: Human-readable metric name
        value:
          type: ["number", "null"]
          description: Raw calculated value
        formattedValue:
          type: ["string", "null"]
          description: Display-formatted value (e.g. "12.4%", "5.2x")
        percentile:
          type: ["number", "null"]
          description: Percentile rank vs S&P500 dataset (0-100)
        color:
          type: string
          enum: [green, amber, red, gray]
          description: Traffic light indicator
        explanation:
          type: ["string", "null"]
          description: One-line explanation of why Buffett cares about this metric
        buffettWhy:
          type: string
          description: Why Buffett weights this metric
        notApplicable:
          type: boolean
          description: True when this metric is excluded for sector reasons

    SimilarStock:
      type: object
      required:
        - ticker
        - name
        - score
        - sector
      properties:
        ticker:
          type: string
        name:
          type: string
        score:
          type: integer
          description: Pre-calculated Buffett score (0-100)
        sector:
          type: string

    StockResult:
      type: object
      required:
        - ticker
        - companyName
        - buffettScore
        - scoreColor
        - scoreConfidence
        - metrics
        - filingDate
        - disclaimer
        - similarStocks
      properties:
        ticker:
          type: string
        companyName:
          type: ["string", "null"]
        sector:
          type: ["string", "null"]
        filingDate:
          type: ["string", "null"]
        buffettScore:
          type: number
          description: Weighted Buffett score 0-100
        scoreColor:
          type: string
          enum: [green, amber, red]
          description: green=70+, amber=40-69, red=<40
        scoreConfidence:
          type: integer
          description: Percentage of applicable metric weight covered by actual data (0-100)
        metrics:
          type: array
          items:
            $ref: "#/components/schemas/MetricScore"
        disclaimer:
          type: string
        similarStocks:
          type: array
          description: 4 stocks from the scored universe whose Buffett scores are closest to this one
          items:
            $ref: "#/components/schemas/SimilarStock"

    StockSummary:
      type: object
      required:
        - ticker
        - companyName
        - buffettScore
        - scoreColor
        - sector
      properties:
        ticker:
          type: string
        companyName:
          type: ["string", "null"]
        sector:
          type: ["string", "null"]
        buffettScore:
          type: number
        scoreColor:
          type: string
          enum: [green, amber, red]