# lib/api-zod/src/generated/types/metricScore.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Buffett's Buffet API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MetricScoreColor } from './metricScoreColor';

export interface MetricScore {
  /** Machine-readable metric key */
  key: string;
  /** Human-readable metric name */
  name: string;
  /**
     * Raw calculated value
     * @nullable
     */
  value: number | null;
  /**
     * Display-formatted value (e.g. "12.4%", "5.2x")
     * @nullable
     */
  formattedValue?: string | null;
  /**
     * Percentile rank vs S&P500 dataset (0-100)
     * @nullable
     */
  percentile: number | null;
  /** Traffic light indicator */
  color: MetricScoreColor;
  /**
     * One-line explanation of why Buffett cares about this metric
     * @nullable
     */
  explanation: string | null;
  /** Why Buffett weights this metric */
  buffettWhy: string;
  /** True when this metric is excluded for sector reasons */
  notApplicable: boolean;
}