# lib/api-zod/src/generated/types/similarStock.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Buffett's Buffet API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SimilarStock {
  ticker: string;
  name: string;
  /** Pre-calculated Buffett score (0-100) */
  score: number;
  sector: string;
}