# lib/api-zod/src/generated/types/stockResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Buffett's Buffet API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MetricScore } from './metricScore';
import type { SimilarStock } from './similarStock';
import type { StockResultScoreColor } from './stockResultScoreColor';

export interface StockResult {
  ticker: string;
  /** @nullable */
  companyName: string | null;
  /** @nullable */
  sector?: string | null;
  /** @nullable */
  filingDate: string | null;
  /** Weighted Buffett score 0-100 */
  buffettScore: number;
  /** green=70+, amber=40-69, red=<40 */
  scoreColor: StockResultScoreColor;
  /** Percentage of applicable metric weight covered by actual data (0-100) */
  scoreConfidence: number;
  metrics: MetricScore[];
  disclaimer: string;
  /** 4 stocks from the scored universe whose Buffett scores are closest to this one */
  similarStocks: SimilarStock[];
}